+++ Packages/com.chark.scriptable-events/Editor/Events/ColorScriptableEventEditor.cs
@@ -1,0 +1,15 @@
+﻿using ScriptableEvents.Events;
+using UnityEditor;
+using UnityEngine;
+
+namespace ScriptableEvents.Editor.Events
+{
+    [CustomEditor(typeof(ColorScriptableEvent))]
+    public class ColorScriptableEventEditor : TypedScriptableEventEditor<Color>
+    {
+        protected override Color DrawArgField(Color value)
+        {
+            return EditorGUILayout.ColorField(value);
+        }
+    }
+}