+++ Packages/com.chark.scriptable-events/Runtime/Events/Collision2DScriptableEvent.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+
+namespace ScriptableEvents.Events
+{
+    [CreateAssetMenu(
+        fileName = "Collision2DScriptableEvent",
+        menuName = "Scriptable Events/Collision 2D Scriptable Event",
+        order = 102
+    )]
+    public class Collision2DScriptableEvent : BaseScriptableEvent<Collision2D>
+    {
+    }
+}