+++ Packages/com.chark.scriptable-events/Runtime/Events/ColorScriptableEvent.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+
+namespace ScriptableEvents.Events
+{
+    [CreateAssetMenu(
+        fileName = "ColorScriptableEvent",
+        menuName = "Scriptable Events/Color Scriptable Event",
+        order = 105
+    )]
+    public class ColorScriptableEvent : BaseScriptableEvent<Color>
+    {
+    }
+}