+++ Packages/com.chark.scriptable-events/Runtime/Events/QuaternionScriptableEvent.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+
+namespace ScriptableEvents.Events
+{
+    [CreateAssetMenu(
+        fileName = "QuaternionScriptableEvent",
+        menuName = "Scriptable Events/Quaternion Scriptable Event",
+        order = 104
+    )]
+    public class QuaternionScriptableEvent : BaseScriptableEvent<Quaternion>
+    {
+    }
+}