+++ Packages/com.chark.scriptable-events/Runtime/Events/Vector2ScriptableEvent.cs
@@ -5,7 +5,7 @@ namespace ScriptableEvents.Events
     [CreateAssetMenu(
         fileName = "Vector2ScriptableEvent",
         menuName = "Scriptable Events/Vector2 Scriptable Event",
-        order = 7
+        order = 100
     )]
     public class Vector2ScriptableEvent : BaseScriptableEvent<Vector2>
     {