+++ Packages/com.chark.scriptable-events/Runtime/Listeners/Collider2DScriptableEventListener.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace ScriptableEvents.Listeners
+{
+    [AddComponentMenu("Scriptable Events/Collider 2D Scriptable Event Listener", 200)]
+    public class Collider2DScriptableEventListener : BaseScriptableEventListener<Collider2D>
+    {
+    }
+}