+++ Packages/com.chark.scriptable-events/Runtime/Listeners/ColliderScriptableEventListener.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace ScriptableEvents.Listeners
+{
+    [AddComponentMenu("Scriptable Events/Collider Scriptable Event Listener", 201)]
+    public class ColliderScriptableEventListener : BaseScriptableEventListener<Collider>
+    {
+    }
+}