+++ Packages/com.chark.scriptable-events/Runtime/Listeners/DoubleScriptableEventListener.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace ScriptableEvents.Listeners
+{
+    [AddComponentMenu("Scriptable Events/Double Scriptable Event Listener", 4)]
+    public class DoubleScriptableEventListener : BaseScriptableEventListener<double>
+    {
+    }
+}