+++ Packages/com.chark.scriptable-events/Runtime/Listeners/FloatScriptableEventListener.cs
@@ -2,7 +2,7 @@
 
 namespace ScriptableEvents.Listeners
 {
-    [AddComponentMenu("Scriptable Events/Float Scriptable Event Listener", 2)]
+    [AddComponentMenu("Scriptable Events/Float Scriptable Event Listener", 3)]
     public class FloatScriptableEventListener : BaseScriptableEventListener<float>
     {
     }