+++ Packages/com.chark.scriptable-events/Runtime/Listeners/LongScriptableEventListener.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace ScriptableEvents.Listeners
+{
+    [AddComponentMenu("Scriptable Events/Long Scriptable Event Listener", 2)]
+    public class LongScriptableEventListener : BaseScriptableEventListener<long>
+    {
+    }
+}