+++ Packages/com.chark.scriptable-events/Runtime/Listeners/TransformScriptableEventListener.cs
@@ -2,7 +2,7 @@
 
 namespace ScriptableEvents.Listeners
 {
-    [AddComponentMenu("Scriptable Events/Transform Scriptable Event Listener", 6)]
+    [AddComponentMenu("Scriptable Events/Transform Scriptable Event Listener", 203)]
     public class TransformScriptableEventListener : BaseScriptableEventListener<Transform>
     {
     }