+++ Packages/com.chark.scriptable-events/Runtime/Listeners/Vector3ScriptableEventListener.cs
@@ -2,7 +2,7 @@
 
 namespace ScriptableEvents.Listeners
 {
-    [AddComponentMenu("Scriptable Events/Vector3 Scriptable Event Listener", 8)]
+    [AddComponentMenu("Scriptable Events/Vector3 Scriptable Event Listener", 101)]
     public class Vector3ScriptableEventListener : BaseScriptableEventListener<Vector3>
     {
     }